Rename pipe files
