+++ src/app/access-requests/features/access-grant-manager-list/access-grant-manager-list.ts
@@ -22,9 +22,9 @@ import { MatSort, MatSortModule } from '@angular/material/sort';
 import { MatTableDataSource, MatTableModule } from '@angular/material/table';
 import { Router, RouterLink } from '@angular/router';
 import { AccessGrant } from '@app/access-requests/models/access-requests';
-import { AccessGrantStatusClassPipe } from '@app/access-requests/pipes/access-grant-status-class.pipe';
+import { AccessGrantStatusClassPipe } from '@app/access-requests/pipes/access-grant-status-class-pipe';
 import { AccessRequestService } from '@app/access-requests/services/access-request';
-import { DatePipe } from '@app/shared/pipes/date.pipe';
+import { DatePipe } from '@app/shared/pipes/date-pipe';
 import {
   DEFAULT_DATE_OUTPUT_FORMAT,
   DEFAULT_TIME_ZONE,
