+++ src/app/access-requests/features/access-request-manager-filter/access-request-manager-filter.ts
@@ -15,7 +15,7 @@ import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
 import { AccessRequestStatus } from '@app/access-requests/models/access-requests';
 import { AccessRequestService } from '@app/access-requests/services/access-request';
-import { Capitalise } from '@app/shared/pipes/capitalise.pipe';
+import { Capitalise } from '@app/shared/pipes/capitalise-pipe';
 import { DATE_INPUT_FORMAT_HINT } from '@app/shared/utils/date-formats';
 
 /**
