+++ src/app/metadata/pipes/facet-activity-pipe.ts
