+++ src/app/metadata/pipes/validate-doi-pipe.ts
