+++ src/app/shared/pipes/add-plural-s-pipe.ts
