+++ src/app/shared/pipes/parse-bytes-pipe.ts
