+++ src/app/shared/pipes/split-lines-pipe.ts
