+++ src/app/shared/pipes/truncate-string-pipe.ts
