+++ src/app/tools/pipes/schemapack-error-pipe.ts
