+++ src/app/verification-addresses/features/new-iva-dialog/new-iva-dialog.ts
@@ -19,7 +19,7 @@ import {
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { IvaType } from '@app/verification-addresses/models/iva';
-import { IvaTypePipe } from '@app/verification-addresses/pipes/iva-type.pipe';
+import { IvaTypePipe } from '@app/verification-addresses/pipes/iva-type-pipe';
 
 /**
  * IVA creation dialog component
