+++ src/app/verification-addresses/features/user-iva-list/user-iva-list.ts
@@ -12,7 +12,7 @@ import { MatTableModule } from '@angular/material/table';
 import { ConfirmationService } from '@app/shared/services/confirmation';
 import { NotificationService } from '@app/shared/services/notification';
 import { Iva, IvaState, IvaType } from '@app/verification-addresses/models/iva';
-import { IvaTypePipe } from '@app/verification-addresses/pipes/iva-type.pipe';
+import { IvaTypePipe } from '@app/verification-addresses/pipes/iva-type-pipe';
 import { IvaService } from '@app/verification-addresses/services/iva';
 import { NewIvaDialogComponent } from '../new-iva-dialog/new-iva-dialog';
 import { VerificationDialogComponent } from '../verification-dialog/verification-dialog';
